Update .NET SDKs/Runtimes, git, git-lfs, kubectl and PowerShell versions
The commit updates the version numbers of the .NET SDKs and Runtimes, git, git-lfs, kubectl and PowerShell in both the README.md and Dockerfile. Moreover, the references to the updated versions' sha512 validation values are modified in the Dockerfile. The URL to download the git-lfs package has also been adjusted according to the new version.