Update .NET SDK, Docker CLI, PowerShell, and kubectl versions
In this commit, the versions of .NET SDK, Docker CLI, PowerShell, and kubectl were updated as mentioned in the README.md file and in the Dockerfile. The .NET SDK 7.0 was also removed from the Dockerfile. The update also includes a change of the image release version in the file 'versions'.